Convert the Trace.session field from PositiveIntegerField to BigIntegerField

+++ lictor/models.py
@@ -3,5 +3,5 @@ from django.db import models
 
 class Trace(models.Model):
     created = models.DateTimeField(auto_now_add=True)
-    session = models.PositiveIntegerField()
+    session = models.BigIntegerField()
     json = models.TextField()
